ai-incident-analyzer/services/alert-receiver/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
kafka-python==2.0.2
